Fix: 결제 검증 api 에러 해결

+++ src/main/java/com/otakumap/domain/payment/dto/PaymentVerifyRequest.java
@@ -3,6 +3,8 @@ package com.otakumap.domain.payment.dto;
 import jakarta.validation.constraints.NotNull;
 import lombok.Getter;
 
+import java.math.BigDecimal;
+
 @Getter
 public class PaymentVerifyRequest {
     @NotNull(message = "imp_uid는 필수입니다.")
@@ -10,5 +12,5 @@ public class PaymentVerifyRequest {
     @NotNull(message = "merchant_uid는 필수입니다.")
     String merchantUid;
     @NotNull(message = "amount는 필수입니다.")
-    Long amount;
+    BigDecimal amount;
 }
